QuantasticaAPI/ai/requirements.txt
fastapi
uvicorn
google-cloud-speech
google-cloud-aiplatform
vertexai
python-dotenv
deprecated
google-adk[database]==0.3.0
poetry
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
pandas
pandas-ta
nsepy
matplotlib
plotly
kaleido
alpaca-py